+++ src/main/java/com/github/liblevenshtein/collection/dictionary/IDawg.java
@@ -1,4 +1,4 @@
-package com.github.liblevenshtein.collection.dawg;
+package com.github.liblevenshtein.collection.dictionary;
 
 import java.io.Serializable;
 import java.util.Collection;