+++ src/main/java/com/github/liblevenshtein/collection/dictionary/factory/IPrefixFactory.java
@@ -1,8 +1,8 @@
-package com.github.liblevenshtein.collection.dawg.factory;
+package com.github.liblevenshtein.collection.dictionary.factory;
 
 import java.io.Serializable;
 
-import com.github.liblevenshtein.collection.dawg.Prefix;
+import com.github.liblevenshtein.collection.dictionary.Prefix;
 
 /**
  * Builds instances of {@link Prefix}.