+++ src/main/java/com/github/liblevenshtein/collection/dictionary/package-info.java
@@ -3,4 +3,4 @@
  * @author Dylon Edwards
  * @since 2.1.0
  */
-package com.github.liblevenshtein.collection.dawg;
+package com.github.liblevenshtein.collection.dictionary;